+++ x-pack/.i18nrc.json
@@ -26,7 +26,7 @@
     "xpack.licensing": "plugins/licensing",
     "xpack.logstash": "legacy/plugins/logstash",
     "xpack.main": "legacy/plugins/xpack_main",
-    "xpack.maps": "legacy/plugins/maps",
+    "xpack.maps": ["plugins/maps", "legacy/plugins/maps"],
     "xpack.ml": ["plugins/ml", "legacy/plugins/ml"],
     "xpack.monitoring": "legacy/plugins/monitoring",
     "xpack.remoteClusters": "plugins/remote_clusters",
