+++ x-pack/legacy/plugins/maps/index.js
@@ -52,7 +52,6 @@ export function maps(kibana) {
         };
       },
       embeddableFactories: ['plugins/maps/embeddable/map_embeddable_factory'],
-      inspectorViews: ['plugins/maps/inspector/views/register_views'],
       home: ['plugins/maps/legacy_register_feature'],
       styleSheetPaths: `${__dirname}/public/index.scss`,
       savedObjectSchemas: {
