+++ x-pack/legacy/plugins/maps/public/actions/ui_actions.js
@@ -4,16 +4,28 @@
  * you may not use this file except in compliance with the Elastic License.
  */
 
-export const UPDATE_FLYOUT = 'UPDATE_FLYOUT';
-export const CLOSE_SET_VIEW = 'CLOSE_SET_VIEW';
-export const OPEN_SET_VIEW = 'OPEN_SET_VIEW';
-export const SET_IS_LAYER_TOC_OPEN = 'SET_IS_LAYER_TOC_OPEN';
-export const SET_FULL_SCREEN = 'SET_FULL_SCREEN';
-export const SET_READ_ONLY = 'SET_READ_ONLY';
-export const SET_OPEN_TOC_DETAILS = 'SET_OPEN_TOC_DETAILS';
-export const SHOW_TOC_DETAILS = 'SHOW_TOC_DETAILS';
-export const HIDE_TOC_DETAILS = 'HIDE_TOC_DETAILS';
-export const UPDATE_INDEXING_STAGE = 'UPDATE_INDEXING_STAGE';
+import {
+  UPDATE_FLYOUT,
+  CLOSE_SET_VIEW,
+  OPEN_SET_VIEW,
+  SET_IS_LAYER_TOC_OPEN,
+  SET_FULL_SCREEN,
+  SET_READ_ONLY,
+  SET_OPEN_TOC_DETAILS,
+  SHOW_TOC_DETAILS,
+  HIDE_TOC_DETAILS,
+  UPDATE_INDEXING_STAGE,
+  // eslint-disable-next-line @kbn/eslint/no-restricted-paths
+} from '../../../../../plugins/maps/public/actions/ui_actions';
+// eslint-disable-next-line @kbn/eslint/no-restricted-paths
+export * from '../../../../../plugins/maps/public/actions/ui_actions';
+
+export function exitFullScreen() {
+  return {
+    type: SET_FULL_SCREEN,
+    isFullScreen: false,
+  };
+}
 
 export function updateFlyout(display) {
   return {
@@ -35,12 +47,6 @@ export function setIsLayerTOCOpen(isLayerTOCOpen) {
   return {
     type: SET_IS_LAYER_TOC_OPEN,
     isLayerTOCOpen,
-  };
-}
-export function exitFullScreen() {
-  return {
-    type: SET_FULL_SCREEN,
-    isFullScreen: false,
   };
 }
 export function enableFullScreen() {
