+++ x-pack/legacy/plugins/maps/public/angular/services/saved_gis_map.js
@@ -18,7 +18,8 @@ import {
 } from '../../selectors/map_selectors';
 import { getIsLayerTOCOpen, getOpenTOCDetails } from '../../selectors/ui_selectors';
 import { convertMapExtentToPolygon } from '../../elasticsearch_geo_utils';
-import { copyPersistentState } from '../../reducers/util';
+// eslint-disable-next-line @kbn/eslint/no-restricted-paths
+import { copyPersistentState } from '../../../../../../plugins/maps/public/reducers/util';
 import { extractReferences, injectReferences } from '../../../common/migrations/references';
 import { MAP_SAVED_OBJECT_TYPE } from '../../../common/constants';
