+++ x-pack/legacy/plugins/maps/public/connected_components/gis_map/index.js
@@ -6,7 +6,8 @@
 
 import { connect } from 'react-redux';
 import { GisMap } from './view';
-import { FLYOUT_STATE } from '../../reducers/ui';
+// eslint-disable-next-line @kbn/eslint/no-restricted-paths
+import { FLYOUT_STATE } from '../../../../../../plugins/maps/public/reducers/ui';
 import { exitFullScreen } from '../../actions/ui_actions';
 import { getFlyoutDisplay, getIsFullScreen } from '../../selectors/ui_selectors';
 import { triggerRefreshTimer, cancelAllInFlightRequests } from '../../actions/map_actions';
