+++ x-pack/legacy/plugins/maps/public/connected_components/layer_addpanel/source_editor/index.js
@@ -6,7 +6,8 @@
 
 import { connect } from 'react-redux';
 import { SourceEditor } from './view';
-import { getInspectorAdapters } from '../../../reducers/non_serializable_instances';
+// eslint-disable-next-line @kbn/eslint/no-restricted-paths
+import { getInspectorAdapters } from '../../../../../../../plugins/maps/public/reducers/non_serializable_instances';
 
 function mapStateToProps(state = {}) {
   return {
