+++ x-pack/legacy/plugins/maps/public/connected_components/layer_panel/flyout_footer/index.js
@@ -6,7 +6,8 @@
 
 import { connect } from 'react-redux';
 import { FlyoutFooter } from './view';
-import { FLYOUT_STATE } from '../../../reducers/ui';
+// eslint-disable-next-line @kbn/eslint/no-restricted-paths
+import { FLYOUT_STATE } from '../../../../../../../plugins/maps/public/reducers/ui';
 import { updateFlyout } from '../../../actions/ui_actions';
 import { hasDirtyState } from '../../../selectors/map_selectors';
 import {
