+++ x-pack/legacy/plugins/maps/public/connected_components/widget_overlay/layer_control/index.js
@@ -6,7 +6,8 @@
 
 import { connect } from 'react-redux';
 import { LayerControl } from './view';
-import { FLYOUT_STATE } from '../../../reducers/ui';
+// eslint-disable-next-line @kbn/eslint/no-restricted-paths
+import { FLYOUT_STATE } from '../../../../../../../plugins/maps/public/reducers/ui.js';
 import { updateFlyout, setIsLayerTOCOpen } from '../../../actions/ui_actions';
 import { setSelectedLayer } from '../../../actions/map_actions';
 import {
