+++ x-pack/legacy/plugins/maps/public/embeddable/map_embeddable.js
@@ -19,7 +19,8 @@ import { esFilters } from '../../../../../../src/plugins/data/public';
 import { I18nContext } from 'ui/i18n';
 
 import { GisMap } from '../connected_components/gis_map';
-import { createMapStore } from '../reducers/store';
+// eslint-disable-next-line @kbn/eslint/no-restricted-paths
+import { createMapStore } from '../../../../../plugins/maps/public/reducers/store';
 import { npStart } from 'ui/new_platform';
 import {
   setGotoWithCenter,
@@ -36,7 +37,11 @@ import {
 } from '../actions/map_actions';
 import { setReadOnly, setIsLayerTOCOpen, setOpenTOCDetails } from '../actions/ui_actions';
 import { getIsLayerTOCOpen, getOpenTOCDetails } from '../selectors/ui_selectors';
-import { getInspectorAdapters, setEventHandlers } from '../reducers/non_serializable_instances';
+import {
+  getInspectorAdapters,
+  setEventHandlers,
+  // eslint-disable-next-line @kbn/eslint/no-restricted-paths
+} from '../../../../../plugins/maps/public/reducers/non_serializable_instances';
 import { getMapCenter, getMapZoom, getHiddenLayerIds } from '../selectors/map_selectors';
 import { MAP_SAVED_OBJECT_TYPE } from '../../common/constants';
