+++ x-pack/legacy/plugins/maps/public/embeddable/merge_input_with_saved_map.js
@@ -5,7 +5,8 @@
  */
 
 import _ from 'lodash';
-import { DEFAULT_IS_LAYER_TOC_OPEN } from '../reducers/ui';
+// eslint-disable-next-line @kbn/eslint/no-restricted-paths
+import { DEFAULT_IS_LAYER_TOC_OPEN } from '../../../../../plugins/maps/public/reducers/ui';
 
 const MAP_EMBEDDABLE_INPUT_KEYS = [
   'hideFilterActions',
