+++ x-pack/legacy/plugins/maps/public/selectors/map_selectors.test.js
@@ -9,7 +9,7 @@ jest.mock('../layers/heatmap_layer', () => {});
 jest.mock('../layers/vector_tile_layer', () => {});
 jest.mock('../layers/sources/all_sources', () => {});
 jest.mock('../layers/joins/inner_join', () => {});
-jest.mock('../reducers/non_serializable_instances', () => ({
+jest.mock('../../../../../plugins/maps/public/reducers/non_serializable_instances', () => ({
   getInspectorAdapters: () => {
     return {};
   },
