+++ x-pack/plugins/maps/common/constants.ts
@@ -1,0 +1,188 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License;
+ * you may not use this file except in compliance with the Elastic License.
+ */
+
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License;
+ * you may not use this file except in compliance with the Elastic License.
+ */
+import { i18n } from '@kbn/i18n';
+export const EMS_CATALOGUE_PATH = 'ems/catalogue';
+
+export const EMS_FILES_CATALOGUE_PATH = 'ems/files';
+export const EMS_FILES_API_PATH = 'ems/files';
+export const EMS_FILES_DEFAULT_JSON_PATH = 'file';
+export const EMS_GLYPHS_PATH = 'fonts';
+export const EMS_SPRITES_PATH = 'sprites';
+
+export const EMS_TILES_CATALOGUE_PATH = 'ems/tiles';
+export const EMS_TILES_API_PATH = 'ems/tiles';
+export const EMS_TILES_RASTER_STYLE_PATH = 'raster/style';
+export const EMS_TILES_RASTER_TILE_PATH = 'raster/tile';
+
+export const EMS_TILES_VECTOR_STYLE_PATH = 'vector/style';
+export const EMS_TILES_VECTOR_SOURCE_PATH = 'vector/source';
+export const EMS_TILES_VECTOR_TILE_PATH = 'vector/tile';
+
+export const MAP_SAVED_OBJECT_TYPE = 'map';
+export const APP_ID = 'maps';
+export const APP_ICON = 'gisApp';
+export const TELEMETRY_TYPE = 'maps-telemetry';
+
+export const MAP_APP_PATH = `app/${APP_ID}`;
+export const GIS_API_PATH = `api/${APP_ID}`;
+export const INDEX_SETTINGS_API_PATH = `${GIS_API_PATH}/indexSettings`;
+
+export const MAP_BASE_URL = `/${MAP_APP_PATH}#/${MAP_SAVED_OBJECT_TYPE}`;
+
+export function createMapPath(id: string) {
+  return `${MAP_BASE_URL}/${id}`;
+}
+
+export const LAYER_TYPE = {
+  TILE: 'TILE',
+  VECTOR: 'VECTOR',
+  VECTOR_TILE: 'VECTOR_TILE',
+  HEATMAP: 'HEATMAP',
+};
+
+export enum SORT_ORDER {
+  ASC = 'asc',
+  DESC = 'desc',
+}
+
+export const EMS_TMS = 'EMS_TMS';
+export const EMS_FILE = 'EMS_FILE';
+export const ES_GEO_GRID = 'ES_GEO_GRID';
+export const ES_SEARCH = 'ES_SEARCH';
+export const ES_PEW_PEW = 'ES_PEW_PEW';
+export const EMS_XYZ = 'EMS_XYZ'; // identifies a custom TMS source. Name is a little unfortunate.
+
+export enum FIELD_ORIGIN {
+  SOURCE = 'source',
+  JOIN = 'join',
+}
+
+export const SOURCE_DATA_ID_ORIGIN = 'source';
+export const META_ID_ORIGIN_SUFFIX = 'meta';
+export const SOURCE_META_ID_ORIGIN = `${SOURCE_DATA_ID_ORIGIN}_${META_ID_ORIGIN_SUFFIX}`;
+export const FORMATTERS_ID_ORIGIN_SUFFIX = 'formatters';
+export const SOURCE_FORMATTERS_ID_ORIGIN = `${SOURCE_DATA_ID_ORIGIN}_${FORMATTERS_ID_ORIGIN_SUFFIX}`;
+
+export const GEOJSON_FILE = 'GEOJSON_FILE';
+
+export const MIN_ZOOM = 0;
+export const MAX_ZOOM = 24;
+
+export const DECIMAL_DEGREES_PRECISION = 5; // meters precision
+export const ZOOM_PRECISION = 2;
+export const DEFAULT_MAX_RESULT_WINDOW = 10000;
+export const DEFAULT_MAX_INNER_RESULT_WINDOW = 100;
+export const DEFAULT_MAX_BUCKETS_LIMIT = 10000;
+
+export const FEATURE_ID_PROPERTY_NAME = '__kbn__feature_id__';
+export const FEATURE_VISIBLE_PROPERTY_NAME = '__kbn_isvisibleduetojoin__';
+
+export const MB_SOURCE_ID_LAYER_ID_PREFIX_DELIMITER = '_';
+
+export const ES_GEO_FIELD_TYPE = {
+  GEO_POINT: 'geo_point',
+  GEO_SHAPE: 'geo_shape',
+};
+
+export const ES_SPATIAL_RELATIONS = {
+  INTERSECTS: 'INTERSECTS',
+  DISJOINT: 'DISJOINT',
+  WITHIN: 'WITHIN',
+};
+
+export const GEO_JSON_TYPE = {
+  POINT: 'Point',
+  MULTI_POINT: 'MultiPoint',
+  LINE_STRING: 'LineString',
+  MULTI_LINE_STRING: 'MultiLineString',
+  POLYGON: 'Polygon',
+  MULTI_POLYGON: 'MultiPolygon',
+  GEOMETRY_COLLECTION: 'GeometryCollection',
+};
+
+export const POLYGON_COORDINATES_EXTERIOR_INDEX = 0;
+export const LON_INDEX = 0;
+export const LAT_INDEX = 1;
+
+export const EMPTY_FEATURE_COLLECTION = {
+  type: 'FeatureCollection',
+  features: [],
+};
+
+export const DRAW_TYPE = {
+  BOUNDS: 'BOUNDS',
+  POLYGON: 'POLYGON',
+};
+
+export enum AGG_TYPE {
+  AVG = 'avg',
+  COUNT = 'count',
+  MAX = 'max',
+  MIN = 'min',
+  SUM = 'sum',
+  TERMS = 'terms',
+  UNIQUE_COUNT = 'cardinality',
+}
+
+export enum RENDER_AS {
+  HEATMAP = 'heatmap',
+  POINT = 'point',
+  GRID = 'grid',
+}
+
+export enum GRID_RESOLUTION {
+  COARSE = 'COARSE',
+  FINE = 'FINE',
+  MOST_FINE = 'MOST_FINE',
+}
+
+export const TOP_TERM_PERCENTAGE_SUFFIX = '__percentage';
+
+export const COUNT_PROP_LABEL = i18n.translate('xpack.maps.aggs.defaultCountLabel', {
+  defaultMessage: 'count',
+});
+
+export const COUNT_PROP_NAME = 'doc_count';
+
+export const STYLE_TYPE = {
+  STATIC: 'STATIC',
+  DYNAMIC: 'DYNAMIC',
+};
+
+export const LAYER_STYLE_TYPE = {
+  VECTOR: 'VECTOR',
+  HEATMAP: 'HEATMAP',
+};
+
+export const COLOR_MAP_TYPE = {
+  CATEGORICAL: 'CATEGORICAL',
+  ORDINAL: 'ORDINAL',
+};
+
+export const COLOR_PALETTE_MAX_SIZE = 10;
+
+export const CATEGORICAL_DATA_TYPES = ['string', 'ip', 'boolean'];
+export const ORDINAL_DATA_TYPES = ['number', 'date'];
+
+export enum SYMBOLIZE_AS_TYPES {
+  CIRCLE = 'circle',
+  ICON = 'icon',
+}
+
+export enum LABEL_BORDER_SIZES {
+  NONE = 'NONE',
+  SMALL = 'SMALL',
+  MEDIUM = 'MEDIUM',
+  LARGE = 'LARGE',
+}
+
+export const DEFAULT_ICON = 'airfield';
