+++ x-pack/plugins/maps/common/parse_xml_string.test.js
