+++ x-pack/plugins/maps/kibana.json
@@ -1,0 +1,8 @@
+{
+  "id": "maps",
+  "version": "8.0.0",
+  "kibanaVersion": "kibana",
+  "configPath": ["xpack", "maps"],
+  "requiredPlugins": ["inspector"],
+  "ui": true
+}
