+++ x-pack/plugins/maps/public/inspector/adapters/map_adapter.js
