+++ x-pack/plugins/maps/public/inspector/views/map_details.js
