+++ x-pack/plugins/maps/public/reducers/map.test.js
