+++ x-pack/plugins/maps/public/reducers/ui.js
