+++ x-pack/plugins/maps/public/reducers/util.js
