+++ x-pack/plugins/maps/public/reducers/util.test.js
